Renamed old TestDataFactory to DeprecatedTestDataFactory
